move useStakeInVault and useClaimVaultReward to API client
