+++ src/clients/api/index.ts
@@ -126,6 +126,10 @@ export { default as swapTokens } from './mutations/swapTokens';
 export * from './mutations/swapTokens';
 export { default as useSwapTokens } from './mutations/swapTokens/useSwapTokens';
 
+export { default as useClaimVaultReward } from './mutations/useClaimVaultReward';
+
+export { default as useStakeInVault } from './mutations/useStakeInVault';
+
 // Queries
 export { default as getVaiTreasuryPercentage } from './queries/getVaiTreasuryPercentage';
 export * from './queries/getVaiTreasuryPercentage';
