+++ src/clients/api/mutations/useClaimVaultReward/index.ts
