--- src/hooks/__mocks__/useClaimVaultReward.ts
@@ -1,7 +1,0 @@
-const claimReward = jest.fn();
-const isLoading = false;
-
-export default () => ({
-  claimReward,
-  isLoading,
-});
