--- src/hooks/__mocks__/useStakeInVault.ts
@@ -1,7 +1,0 @@
-const stake = jest.fn();
-const isLoading = false;
-
-export default () => ({
-  stake,
-  isLoading,
-});
