+++ src/pages/Vault/VaultItem/index.tsx
@@ -12,11 +12,10 @@ import { Token } from 'types';
 import { areTokensEqual, convertWeiToTokens, formatToReadablePercentage } from 'utilities';
 import type { TransactionReceipt } from 'web3-core/types';
 
-import { useWithdrawFromVrtVault } from 'clients/api';
+import { useClaimVaultReward, useWithdrawFromVrtVault } from 'clients/api';
 import { TOKENS } from 'constants/tokens';
 import { AuthContext } from 'context/AuthContext';
 import { DisableLunaUstWarningContext } from 'context/DisableLunaUstWarning';
-import useClaimVaultReward from 'hooks/useClaimVaultReward';
 import useConvertWeiToReadableTokenString from 'hooks/useConvertWeiToReadableTokenString';
 import useHandleTransactionMutation from 'hooks/useHandleTransactionMutation';
